// src/compat/types.ts
import type { Ref } from 'react';
import type { StyleProp, ViewStyle } from 'react-native';
import type { PointGroup } from './signaturePadData';

export type SignatureViewRef = {
  readSignature(): void;
  clearSignature(): void;
  undo(): void;
  redo(): void;
  draw(): void;
  erase(): void;
  changePenColor(color: string): void;
  changePenSize(minWidth: number, maxWidth: number): void;
  getData(): void;
  fromData(pointGroups: PointGroup[], suppressClear?: boolean): void;
  setDataURL(url: string): void;
  reinitialize(): void;
};

export type SignatureCanvasProps = {
  ref?: Ref<SignatureViewRef>;
  testID?: string;
  style?: StyleProp<ViewStyle>;
  penColor?: string;
  minWidth?: number;
  maxWidth?: number;
  dotSize?: number;
  backgroundColor?: string;
  bgSrc?: string;
  bgWidth?: number;
  bgHeight?: number;
  dataURL?: string;
  overlaySrc?: string;
  overlayWidth?: number;
  overlayHeight?: number;
  trimWhitespace?: boolean;
  imageType?: 'image/png' | 'image/jpeg' | 'image/svg+xml';
  autoClear?: boolean;
  descriptionText?: string;
  clearText?: string;
  confirmText?: string;
  webStyle?: string;
  onOK?: (signature: string) => void;
  onEmpty?: () => void;
  onClear?: () => void;
  onBegin?: () => void;
  onEnd?: () => void;
  onUndo?: () => void;
  onRedo?: () => void;
  onDraw?: () => void;
  onErase?: () => void;
  onChangePenColor?: () => void;
  onChangePenSize?: () => void;
  onGetData?: (data: string) => void;
  onLoadEnd?: () => void;
  onError?: (error: Error) => void;
  customHtml?: unknown;
  webviewProps?: unknown;
  webviewContainerStyle?: unknown;
  androidLayerType?: unknown;
  androidHardwareAccelerationDisabled?: boolean;
  nestedScrollEnabled?: boolean;
  showsVerticalScrollIndicator?: boolean;
  scrollable?: boolean;
  rotated?: boolean;
  velocityFilterWeight?: number;
};